library/axi_pwm_custom: Add PWM IP
Declare LED PWM output signals ('pwm_led_0'...)
Added algorithm to generate PWM proportional to the input signals.
The code also samples the input signal and holds it for an entire period, so that PWM can be generated.
Added a variable (namely 'inc_rate') to check the functionality of the above mentioned sample and hold.