library/axi_pwm_custum: Add IP for PWM
In this project we have the ADXL327 accelerometer connected to the FPGA.
This is communicating with SPI. Depending on how much is the
accelerometer tilted in the positive or negative direction of the X and
Y axes respectively, the MISO will return a value. Maximum tilt is
equivalent to 2^12-1 on MISO. These values can be checked with a logic
analyzer connected to the JD PMOD, but we want to interface the value
with LEDs too.

The PWM is used to adjust the brightness of 4 LEDs. Each LED corresponds
to an ax of the accelerometer.

Signed-off-by: Daniel Matyas <daniel.matyas@analog.com>