library/verilog_task_testbench: Created a module and a testbench for it
The main target was to create a module that generates a saw-tooth like signal
and to create a testbench for the module, in order to see what I wanted to
acomplish.
-The main module(verilog_task_testbench.v) contains a 12b counter that goes up to
it's maximum value, counts back to zero and then it goes back, on repeat. We use this
counter to generate a triangle like waveform(in analog waveform representaion).
-Imported the module in the testbench. In here, we emulate the clock and reset signal