library/axi_pwm_custom: Completed the custom PWM module and the testbench to show it's functionallity
Things that were added/modified:
-added: completed the interface(axi_pwm_custom_if) for the custom PWM module. This includes a counter
that it's used for dividing the clock. The counter is used for controlling the pwm signal based on the inputs recived
from the data channels. Data is processed at a fix period of time, with the help of the clock divider.
-modified: changed some things in the testbench(axi_pwm_custom_if_tb) to show how pwm signals change based
on different input on the data channels.
-modified : in PWM top module(axi_pwm_custom), completed the missing signals and added the custom interface
that was just added.