library/axi_pwm_custom : IP for custom PWM generator
Interface for custom PWM generator
The duration of the PWM signal is determined by comparing an input value to a sawtooth signal.
Sawtooth signal is created by incrementing a 12-bit value to maximum than reseting it to 0.
The PWM signal pulse lasts until the value of the sawtooth signal is greater than the input value.
The PWM signal controls the brightness of four LEDs.